+++ Makefile.sdrplay
@@ -1,0 +1,153 @@
+# Primary Linux makefile for ka9q-radio package
+# Copyright 2017-2024, Phil Karn, KA9Q
+
+# for production
+DOPTS = -DNDEBUG=1 -O3
+
+# for debugging
+#DOPTS=-g
+
+# do NOT set -ffast-math or -ffinite-math-only; NANs are widely used as 'variable not set' sentinels
+COPTS=-march=native -std=gnu11 -pthread -Wall -funsafe-math-optimizations -fno-math-errno -fcx-limited-range -D_GNU_SOURCE=1
+
+CFLAGS=$(DOPTS) $(COPTS) $(INCLUDES)
+BINDIR=/usr/local/bin
+LIBDIR=/usr/local/share/ka9q-radio
+DAEMONDIR=/usr/local/sbin
+VARDIR=/var/lib/ka9q-radio
+LDLIBS=-lpthread -lbsd -lm
+
+DAEMONS=aprs aprsfeed cwd opusd packetd radiod stereod rdsd
+
+EXECS=control jt-decoded metadump monitor opussend pcmcat pcmrecord pcmsend pcmspawn pl powers setfilt show-pkt show-sig tune wd-record
+
+
+LOGROTATE_FILES = aprsfeed.rotate ft8.rotate ft4.rotate wspr.rotate
+
+CFILES = aprs.c aprsfeed.c attr.c audio.c avahi.c avahi_browse.c ax25.c bandplan.c config.c control.c cwd.c decimate.c decode_status.c dump.c filter.c fm.c iir.c jt-decoded.c linear.c main.c metadump.c misc.c modes.c monitor.c morse.c multicast.c opusd.c opussend.c osc.c packetd.c pcmcat.c pcmrecord.c pcmsend.c pcmspawn.c pl.c powers.c radio.c radio_status.c rdsd.c rtcp.c sdrplay.c setfilt.c show-pkt.c show-sig.c sig_gen.c spectrum.c status.c stereod.c tune.c wd-record.c wfm.c
+
+HFILES = attr.h ax25.h bandplan.h conf.h config.h decimate.h filter.h iir.h misc.h morse.h multicast.h osc.h radio.h status.h
+
+all: $(DAEMONS) $(EXECS)
+
+install: $(DAEMONS) $(EXECS)
+	install -o root -m 0755 -D --target-directory=$(DAEMONDIR) $(DAEMONS)
+	install -o root -m 0755 -D --target-directory=$(BINDIR) $(EXECS) start-ka9q-horus.sh
+
+# NB: now overwrites existing .conf files in /etc/radio !
+install.phil: $(DAEMONS) $(EXECS)
+	/usr/sbin/adduser --quiet --system --group radio
+	/usr/sbin/adduser --quiet --system --ingroup radio recordings
+	/usr/sbin/adduser --quiet --system --ingroup radio aprsfeed
+	rsync -a aux/98-sockbuf.conf /etc/sysctl.d
+	sysctl --quiet -p /etc/sysctl.d/98-sockbuf.conf
+	rsync -a $(DAEMONS) $(DAEMONDIR)
+	rsync -a $(EXECS) start-ka9q-horus.sh $(BINDIR)
+	setcap cap_sys_nice=ep $(BINDIR)/monitor
+	(cd share;rsync -a --ignore-existing . $(LIBDIR))
+	(cd service;rsync -a . /etc/systemd/system)
+	(cd rules;rsync -a . /etc/udev/rules.d)
+	(cd aux;rsync -a $(BLACKLIST) /etc/modprobe.d)
+	(cd aux;rsync -a $(LOGROTATE_FILES) /etc/logrotate.d;chown root:root /etc/logrotate.d/*)
+	mkdir -p /etc/fftw /etc/radio $(VARDIR)
+	chgrp radio $(VARDIR) /etc/radio /etc/fftw
+	chmod g+ws $(VARDIR) /etc/radio /etc/fftw
+	systemctl daemon-reload
+
+clean:
+	-rm -f *.o *.a .depend $(EXECS) $(DAEMONS)
+
+.PHONY: clean all install
+
+ifeq (,$(findstring $(MAKECMDGOALS),clean))
+     -include .depend
+endif
+
+.depend: $(CFILES) $(HFILES)
+	rm -f .depend
+	$(CC) $(CFLAGS) -MM $(CFILES) > .depend
+
+# Executables
+set_xcvr: set_xcvr.o
+	$(CC) $(LDOPTS) -o $@ $^ -lpigpio -liniparser -lrt -lbsd -lm -lpthread
+
+aprs: aprs.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+aprsfeed: aprsfeed.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+control: control.o bandplan.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lavahi-client -lavahi-common -lfftw3f_threads -lfftw3f -lncursesw -liniparser -lbsd -lm -lpthread
+
+cwd: cwd.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+tune: tune.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+setfilt: setfilt.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+show-pkt: show-pkt.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lncursesw -lbsd -lm -lpthread
+
+show-sig: show-sig.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lncursesw -lbsd -lm -lpthread
+
+metadump: metadump.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+monitor: monitor.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lopus -lportaudio -lasound -lncursesw -liniparser -lbsd -lm -lpthread
+
+opusd: opusd.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lopus -lavahi-client -lavahi-common -lbsd -lm -lpthread
+
+opussend: opussend.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lopus -lportaudio -lbsd -lm -lpthread
+
+packetd: packetd.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lavahi-client -lavahi-common -lfftw3f_threads -lfftw3f -lbsd -lm -lpthread
+
+pcmcat: pcmcat.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+pcmspawn: pcmspawn.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+pcmrecord: pcmrecord.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+pcmsend: pcmsend.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lportaudio -lbsd -lm -lpthread
+
+powers: powers.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+pl: pl.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lfftw3f_threads -lfftw3f -lbsd -lm -lpthread
+
+radiod: main.o audio.o fm.o wfm.o linear.o spectrum.o radio.o radio_status.o rtcp.o sdrplay.o sig_gen.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lavahi-client -lavahi-common -lfftw3f_threads -lfftw3f -liniparser -lsdrplay_api -lportaudio -lbsd -lm -lpthread
+
+rdsd: rdsd.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lavahi-client -lavahi-common -lfftw3f_threads -lfftw3f -lbsd -lm -lpthread
+
+stereod: stereod.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lavahi-client -lavahi-common -lfftw3f_threads -lfftw3f -lbsd -lm -lpthread
+
+jt-decoded: jt-decoded.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lfftw3f_threads -lfftw3f -lbsd -lm -lpthread
+
+wd-record: wd-record.o libradio.a
+	$(CC) $(LDOPTS) -o $@ $^ -lbsd -lm -lpthread
+
+
+# subroutines useful in more than one program
+libradio.a: morse.o dump.o modes.o ax25.o avahi.o avahi_browse.o attr.o filter.o iir.o decode_status.o status.o misc.o multicast.o osc.o config.o
+	ar rv $@ $?
+	ranlib $@
+
+
+
