--- aux/sdrplayd.conf
@@ -1,14 +1,0 @@
-[rspdx-antC]
-description = RSPdx - Antenna C
-serial = 00000000000
-antenna = Antenna C
-#samprate = 2000000        ; default is 2M
-lna-state = 3
-if-agc = yes
-#if-att = 40
-#iface = eth0	           ; force primary interface, avoid wifi
-iface = lo	           ; default loopback interface, avoid wifi
-status = rspdx-antC.local
-data = rspdx-antC-data.local
-
-frequency = 90400000       ; locks tuner when manually set
