+++ config/radiod@sdrplay-generic.conf
@@ -1,0 +1,22 @@
+# allowable Opus block times: 2.5, 5, 10, 20, 40, 60, 80, 100, 120
+[global]
+status = nws.local
+hardware = sdrplay
+iface = lo
+#verbose = 1
+
+[sdrplay]
+device = sdrplay
+description = RSPdx - Antenna C
+#serial =
+antenna = Antenna C
+#samprate = 2000000        ; default is 2M
+lna-state = 2
+if-agc = yes
+
+frequency = 162480000
+
+[NWS]
+mode = fm
+data = nws-pcm.local
+freq = "162m400 162m425 162m450 162m475 162m500 162m525 162m550"
