+++ config/radiod@sdrplay-iq.conf
@@ -1,0 +1,21 @@
+[global]
+status = nws-iq.local
+hardware = sdrplay
+iface = lo
+#verbose = 1
+
+[sdrplay]
+device = sdrplay
+description = RSPdx - Antenna C
+#serial =
+antenna = Antenna C
+#samprate = 2000000        ; default is 2M
+lna-state = 2
+if-agc = yes
+
+frequency = 162480000
+
+[NWS-IQ]
+mode = iq
+data = nws-iq-pcm.local
+freq = 162m550
