+++ config/radiod@sdrplay-wspr-20m.conf
@@ -1,0 +1,22 @@
+# allowable Opus block times: 2.5, 5, 10, 20, 40, 60, 80, 100, 120
+[global]
+status = wspr.local
+hardware = sdrplay
+iface = lo
+#verbose = 1
+
+[sdrplay]
+device = sdrplay
+description = RSPdx - Antenna C
+#serial =
+antenna = Antenna C
+#samprate = 2000000        ; default is 2M
+lna-state = 2
+if-agc = yes
+
+frequency = 14100000
+
+[WSPR]
+mode = usb
+data = wspr-pcm.local
+freq = "136k000 474k200 1m836600 3m592600 5m287200 7m038600 10m138700 14m095600 18m104600 21m094600 24m924600 28m124600"
