+++ main.c
@@ -601,6 +601,7 @@ static int setup_hardware(char const *sname){
   }
   // Do we support it?
   // This should go into a table somewhere
+#if 0
   if(strcasecmp(device,"rx888") == 0){
     Frontend.setup = rx888_setup;
     Frontend.start = rx888_startup;
@@ -622,10 +623,11 @@ static int setup_hardware(char const *sname){
     Frontend.start = rtlsdr_startup;
     Frontend.tune = rtlsdr_tune;
   } else if(strcasecmp(device,"sig_gen") == 0){
+#else
+  if(strcasecmp(device,"sig_gen") == 0){
     Frontend.setup = sig_gen_setup;
     Frontend.start = sig_gen_startup;
     Frontend.tune = sig_gen_tune;
-#if 0
     // The sdrplay library is still proprietary and object-only, so I can't bundle it in ka9q-radio
     // Everything else either has a standard Debian package or I have information to program them directly.
     // To hell with vendors who deliberately make their products hard to use when they have plenty of competition.
